// src/lib/shopify/types.ts
export type Money = {
  amount: string;
  currencyCode: string;
};
// ...
export type ProductOption = {
  id: string;
  name: string;
  values: string[];
};

export type ProductVariant = {
  id: string;
  title: string;
  availableForSale: boolean;
  selectedOptions: {
    name: string;
    value: string;
  }[];
  price: Money;
};